Distance, Kernel Matrices and k-Nearest Neighbours

Compute distance matrices (similarity matrices), convert them into kernel matrices or k-nearest neighbor objects.

Functions

MultivariateAnomalies.dist_matrix — Function.

dist_matrix{tp, N}(data::AbstractArray{tp, N}; dist::ASCIIString = "Euclidean", space::Int = 0, lat::Int = 0, lon::Int = 0, Q = 0)
dist_matrix{tp, N}(data::AbstractArray{tp, N}, training_data; dist::ASCIIString = "Euclidean", space::Int = 0, lat::Int = 0, lon::Int = 0, Q = 0)

compute the distance matrix of data or the distance matrix between data and training data i.e. the pairwise distances along the first dimension of data, using the last dimension as variables. dist is a distance metric, currently Euclidean(default), SqEuclidean, Chebyshev, Cityblock, JSDivergence, Mahalanobis and SqMahalanobis are supported. The latter two need a covariance matrix Q as input argument.

Examples

julia> dc = randn(10, 4,3)
julia> D = dist_matrix(dc, space = 2)

MultivariateAnomalies.dist_matrix! — Function.

dist_matrix!(D_out, data, ...)

compute the distance matrix of data, similar to dist_matrix(). D_out object has to be preallocated, i.e. with init_dist_matrix.

julia> dc = randn(10,4, 4,3)
julia> D_out = init_dist_matrix(dc)
julia> dist_matrix!(D_out, dc, lat = 2, lon = 2)
julia> D_out[1]

MultivariateAnomalies.init_dist_matrix — Function.

init_dist_matrix(data)
init_dist_matrix(data, training_data)

initialize a D_out object for dist_matrix!().

MultivariateAnomalies.knn_dists — Function.

knn_dists(D, k::Int, temp_excl::Int = 5)

returns the k-nearest neighbors of a distance matrix D. Excludes temp_excl (default: temp_excl = 5) distances from the main diagonal of D to be also nearest neighbors.

julia> dc = randn(20, 4,3)
julia> D = dist_matrix(dc, space = 2)
julia> knn_dists_out = knn_dists(D, 3, 1)
julia> knn_dists_out[5] # distances
julia> knn_dists_out[4] # indices

MultivariateAnomalies.knn_dists! — Function.

knn_dists!(knn_dists_out, D, temp_excl::Int = 5)

returns the k-nearest neighbors of a distance matrix D. Similar to knn_dists(), but uses preallocated input object knn_dists_out, initialized with init_knn_dists(). Please note that the number of nearest neighbors k is not necessary, as it is already determined by the knn_dists_out object.

julia> dc = randn(20, 4,3)
julia> D = dist_matrix(dc, space = 2)
julia> knn_dists_out = init_knn_dists(dc, 3)
julia> knn_dists!(knn_dists_out, D)
julia> knn_dists_out[5] # distances
julia> knn_dists_out[4] # indices

MultivariateAnomalies.init_knn_dists — Function.

init_knn_dists(T::Int, k::Int)
init_knn_dists(datacube::AbstractArray, k::Int)

initialize a preallocated knn_dists_out object. kis the number of nerarest neighbors, T the number of time steps (i.e. size of the first dimension) or a multidimensional datacube.

MultivariateAnomalies.kernel_matrix — Function.

kernel_matrix(D::AbstractArray, σ::Float64 = 1.0[, kernel::ASCIIString = "gauss", dimension::Int64 = 1])

compute a kernel matrix out of distance matrix D, given σ. Optionally normalized by the dimension, if kernel = "normalized_gauss". compute D with dist_matrix().

julia> dc = randn(20, 4,3)
julia> D = dist_matrix(dc, space = 2)
julia> K = kernel_matrix(D, 2.0)

MultivariateAnomalies.kernel_matrix! — Function.

kernel_matrix!(K, D::AbstractArray, σ::Float64 = 1.0[, kernel::ASCIIString = "gauss", dimension::Int64 = 1])

compute a kernel matrix out of distance matrix D. Similar to kernel_matrix(), but with preallocated Array K (K = similar(D)) for output.

julia> dc = randn(20, 4,3)
julia> D = dist_matrix(dc, space = 2)
julia> kernel_matrix!(D, D, 2.0) # overwrites distance matrix

Index

MultivariateAnomalies.EWMA

MultivariateAnomalies.EWMA!

MultivariateAnomalies.KDE

MultivariateAnomalies.KDE!

MultivariateAnomalies.KNFST_predict

MultivariateAnomalies.KNFST_predict!

MultivariateAnomalies.KNFST_train

MultivariateAnomalies.KNN_Delta

MultivariateAnomalies.KNN_Delta!

MultivariateAnomalies.KNN_Gamma

MultivariateAnomalies.KNN_Gamma!

MultivariateAnomalies.REC

MultivariateAnomalies.REC!

MultivariateAnomalies.SVDD_predict

MultivariateAnomalies.SVDD_predict!

MultivariateAnomalies.SVDD_train

MultivariateAnomalies.T2

MultivariateAnomalies.T2!

MultivariateAnomalies.TDE

MultivariateAnomalies.UNIV

MultivariateAnomalies.UNIV!

MultivariateAnomalies.auc

MultivariateAnomalies.auc_fpr_tpr

MultivariateAnomalies.boolevents

MultivariateAnomalies.compute_ensemble

MultivariateAnomalies.detectAnomalies

MultivariateAnomalies.detectAnomalies!

MultivariateAnomalies.dist_matrix

MultivariateAnomalies.dist_matrix!

MultivariateAnomalies.getParameters

MultivariateAnomalies.get_MedianCycle

MultivariateAnomalies.get_MedianCycle!

MultivariateAnomalies.get_MedianCycles

MultivariateAnomalies.get_quantile_scores

MultivariateAnomalies.get_quantile_scores!

MultivariateAnomalies.globalICA

MultivariateAnomalies.globalPCA

MultivariateAnomalies.init_KDE

MultivariateAnomalies.init_KNFST

MultivariateAnomalies.init_KNN_Delta

MultivariateAnomalies.init_KNN_Gamma

MultivariateAnomalies.init_MedianCycle

MultivariateAnomalies.init_REC

MultivariateAnomalies.init_SVDD_predict

MultivariateAnomalies.init_T2

MultivariateAnomalies.init_UNIV

MultivariateAnomalies.init_detectAnomalies

MultivariateAnomalies.init_dist_matrix

MultivariateAnomalies.init_knn_dists

MultivariateAnomalies.kernel_matrix

MultivariateAnomalies.kernel_matrix!

MultivariateAnomalies.knn_dists

MultivariateAnomalies.knn_dists!

MultivariateAnomalies.mw_COR

MultivariateAnomalies.mw_VAR

MultivariateAnomalies.sMSC