Scores

Postprocess your anomaly scores by making different algorithms comparable and computing their ensemble.

Functions

MultivariateAnomalies.get_quantile_scores — Function.

get_quantile_scores(scores, quantiles = 0.0:0.01:1.0)

return the quantiles of the given N dimensional anomaly scores cube. quantiles (default: quantiles = 0.0:0.01:1.0) is a Float range of quantiles. Any score being greater or equal quantiles[i] and beeing smaller than quantiles[i+1] is assigned to the respective quantile quantiles[i].

Examples

julia> scores1 = rand(10, 2)
julia> quantile_scores1 = get_quantile_scores(scores1)

MultivariateAnomalies.get_quantile_scores! — Function.

get_quantile_scores!{tp,N}(quantile_scores::AbstractArray{Float64, N}, scores::AbstractArray{tp,N}, quantiles::FloatRange{Float64} = 0.0:0.01:1.0)

return the quantiles of the given N dimensional scores array into a preallocated quantile_scores array, see get_quantile_scores().

MultivariateAnomalies.compute_ensemble — Function.

compute_ensemble(m1_scores, m2_scores[, m3_scores, m4_scores]; ensemble = "mean")

compute the mean (ensemble = "mean"), minimum (ensemble = "min"), maximum (ensemble = "max") or median (ensemble = "median") of the given anomaly scores. Supports between 2 and 4 scores input arrays (m1_scores, ..., m4_scores). The scores of the different anomaly detection algorithms should be somehow comparable, e.g., by using get_quantile_scores() before.

Examples

julia> scores1 = rand(10, 2)
julia> scores2 = rand(10, 2)
julia> quantile_scores1 = get_quantile_scores(scores1)
julia> quantile_scores2 = get_quantile_scores(scores2)
julia> compute_ensemble(quantile_scores1, quantile_scores2, ensemble = "max")

Index

MultivariateAnomalies.EWMA

MultivariateAnomalies.EWMA!

MultivariateAnomalies.KDE

MultivariateAnomalies.KDE!

MultivariateAnomalies.KNFST_predict

MultivariateAnomalies.KNFST_predict!

MultivariateAnomalies.KNFST_train

MultivariateAnomalies.KNN_Delta

MultivariateAnomalies.KNN_Delta!

MultivariateAnomalies.KNN_Gamma

MultivariateAnomalies.KNN_Gamma!

MultivariateAnomalies.REC

MultivariateAnomalies.REC!

MultivariateAnomalies.SVDD_predict

MultivariateAnomalies.SVDD_predict!

MultivariateAnomalies.SVDD_train

MultivariateAnomalies.T2

MultivariateAnomalies.T2!

MultivariateAnomalies.TDE

MultivariateAnomalies.UNIV

MultivariateAnomalies.UNIV!

MultivariateAnomalies.auc

MultivariateAnomalies.auc_fpr_tpr

MultivariateAnomalies.boolevents

MultivariateAnomalies.compute_ensemble

MultivariateAnomalies.detectAnomalies

MultivariateAnomalies.detectAnomalies!

MultivariateAnomalies.dist_matrix

MultivariateAnomalies.dist_matrix!

MultivariateAnomalies.getParameters

MultivariateAnomalies.get_MedianCycle

MultivariateAnomalies.get_MedianCycle!

MultivariateAnomalies.get_MedianCycles

MultivariateAnomalies.get_quantile_scores

MultivariateAnomalies.get_quantile_scores!

MultivariateAnomalies.globalICA

MultivariateAnomalies.globalPCA

MultivariateAnomalies.init_KDE

MultivariateAnomalies.init_KNFST

MultivariateAnomalies.init_KNN_Delta

MultivariateAnomalies.init_KNN_Gamma

MultivariateAnomalies.init_MedianCycle

MultivariateAnomalies.init_REC

MultivariateAnomalies.init_SVDD_predict

MultivariateAnomalies.init_T2

MultivariateAnomalies.init_UNIV

MultivariateAnomalies.init_detectAnomalies

MultivariateAnomalies.init_dist_matrix

MultivariateAnomalies.init_knn_dists

MultivariateAnomalies.kernel_matrix

MultivariateAnomalies.kernel_matrix!

MultivariateAnomalies.knn_dists

MultivariateAnomalies.knn_dists!

MultivariateAnomalies.mw_COR

MultivariateAnomalies.mw_VAR

MultivariateAnomalies.sMSC